Staff user with "view" permissions only, who is part of the workflow, can view drafts and moderate but not edit.
- added user_part_of_moderation method to PageModerationRequest model
to check whether given user is a required approver for this Page’s
moderation Workflow
- added monkeypatch page_permissions.user_can_view_page_draft to allow
any user that is part of the Workflow to view draft content. This calls
the new model method above.
- cms_toolbars:: user_can_publish override was added to show moderation
button if there is a moderation request and the given user is a
required approver of that Workflow